requirements.txt
anthropic==0.57.1
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-tavily==0.2.8
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
openai==1.96.1
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==1.0.0
tiktoken==0.9.0
tqdm==4.67.1
